Split NumberAndTypeConversionTest into two files
This way we avoid the nested classes and experimental Enclosed feature.
